requirements.txt
arcade==1.3.4
future==0.16.0
Pillow==5.1.0
pyglet==1.3.2
aiohttp==3.3.2
pyzmq==17.1.2
pymunk==5.3.2
dominate==2.3.1
